[ALPS04589945] ion: fix build error
[Detail]
fix ion build error
test: build ok

MTK-Commit-Id: d1d27d3740c18be415e46fd55c75590ccf84ffec

CR-Id: ALPS04589945
Feature: [Module]ION/M4U

+++ drivers/misc/mediatek/pseudo_m4u/pseudo_m4u.h
@@ -283,7 +283,7 @@ int pseudo_config_port_tee(int kernelport);
 
 int m4u_mva_map_kernel(unsigned int mva,
 	unsigned long size, unsigned long *map_va,
-	unsigned long *map_size);
+	unsigned int *map_size);
 
 int m4u_mva_unmap_kernel(unsigned int mva,
 		unsigned long size, unsigned long map_va);
